Lots of big UI changes
Also implementing graph-trace on the glsl side but it wasn't as stable as ray-trace. I think this is because floating point numbers get very large for the pixel point we are trying to get to. In ray-trace the positions are never very large - going a long distance is measured by dist_to_go rather than a distant point in space.